@@ -8,26 +8,25 @@
     "build": "vite build",
     "preview": "vite preview",
     "check": "svelte-kit sync && svelte-check --no-tsconfig",
-    "check:watch": "svelte-kit sync && svelte-check --no-tsconfig --watch",
-    "clean": "rimraf build .svelte-kit"
+    "check:watch": "svelte-kit sync && svelte-check --no-tsconfig --watch"
   },
   "dependencies": {
     "@minvws/manon": "workspace:*",
     "@minvws/manon-themes": "workspace:*",
-    "highlight.js": "^11.9.0"
+    "highlight.js": "^11.11.1"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "^3.0.2",
-    "@sveltejs/kit": "^2.5.18",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
-    "prettier": "^3.3.2",
-    "prettier-plugin-svelte": "^3.2.5",
-    "rimraf": "^5.0.8",
-    "sass": "^1.77.6",
-    "svelte": "^4.2.18",
-    "svelte-check": "^3.8.4",
+    "@sveltejs/adapter-static": "^3.0.8",
+    "@sveltejs/kit": "^2.21.4",
+    "@sveltejs/vite-plugin-svelte": "^5.1.0",
+    "mdsvex": "^0.12.6",
+    "prettier": "^3.5.3",
+    "prettier-plugin-svelte": "^3.4.0",
+    "sass": "^1.89.2",
+    "svelte": "^5.33.19",
+    "svelte-check": "^4.2.1",
     "svelte-preprocess-import-assets": "^1.1.0",
-    "typescript": "^5.5.3",
-    "vite": "^5.3.3"
+    "typescript": "^5.8.3",
+    "vite": "^6.3.5"
   }
 }
