@@ -1,7 +1,9 @@
-<script>
+<script lang="ts">
+  import type { HTMLAttributes } from "svelte/elements";
   import { onMount } from "svelte";
-  import { browser } from "$app/environment";
-  import { initSidemenus } from "$lib/manon.js";
+  import { initSidemenus } from "$lib/manon";
+
+  let props: HTMLAttributes<HTMLElement> = $props();
   onMount(initSidemenus);
 </script>
 
@@ -9,7 +11,7 @@
   data-open-label="Zijbalknavigatie"
   data-close-label="Sluit zijbalknavigatie"
   aria-label="Document-navigatie"
-  {...$$props}
+  {...props}
 >
-  <slot />
+  {@render props.children?.()}
 </nav>
