@@ -1,9 +1,12 @@
-<script>
+<script lang="ts">
+  import type { HTMLAttributes } from "svelte/elements";
   import { base } from "$app/paths";
   import NavLink from "$lib/NavLink.svelte";
+
+  let props: HTMLAttributes<HTMLElement> = $props();
 </script>
 
-<ul {...$$props}>
+<ul {...props}>
   <li><NavLink href="{base}/">Home</NavLink></li>
   <li><NavLink href="{base}/components">Componenten</NavLink></li>
   <li><NavLink href="{base}/documentation">Documentatie</NavLink></li>
