@@ -1,10 +1,13 @@
+import { fileURLToPath } from "node:url";
 import adapter from "@sveltejs/adapter-static";
 import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
 import { importAssets } from "svelte-preprocess-import-assets";
+import { mdsvex } from "mdsvex";
 import { redirects } from "./src/redirects.js";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
+  extensions: [".svelte", ".md"],
   kit: {
     adapter: adapter(),
     prerender: {
@@ -19,7 +22,19 @@ const config = {
       relative: true,
     },
   },
-  preprocess: [vitePreprocess(), importAssets()],
+  preprocess: [
+    vitePreprocess(),
+    mdsvex({
+      extensions: [".md"],
+      layout: {
+        _: fileURLToPath(import.meta.resolve("./src/markdown/layouts/default.svelte")),
+      },
+      highlight: {
+        highlighter: (code, lang) => `<Components.Code code={\`${code}\`} language="${lang}" />`,
+      },
+    }),
+    importAssets(),
+  ],
 };
 
 export default config;
